@@ -2,7 +2,7 @@
   "devDependencies": {
     "cssnano": "^4.1.10",
     "parcel-bundler": "^1.12.3",
-    "typescript": "^3.4.5"
+    "typescript": "^3.8.3"
   },
   "name": "frontend",
   "version": "1.0.0",
